refactor: move landing page components to containers directory
