+++ frontend/src/containers/landing-page/HowAndWhy/HowRepairCoinWorks.tsx
